requirements.txt
absl-py==1.4.0
asttokens==2.4.1
attrs==23.2.0
Augmentor==0.2.12
backcall==0.2.0
bagnets @ git+https://github.com/wielandbrendel/bag-of-local-features-models.git@8493fb442c9b9ccfeaa970c950ca67832017f072
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
fonttools==4.39.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
idna==3.4
imageio==2.33.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
lazy_loader==0.3
lit==15.0.7
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.2.1
munch==2.5.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.23.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.2
referencing==0.33.0
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
sklearn==0.0.post5
soupsieve==2.5
stack-data==0.6.3
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tifffile==2023.7.10
tinycss2==1.2.1
torch==2.0.0+cu118
torch-tb-profiler==0.4.1
torchvision==0.15.1+cu118
tornado==6.4
tqdm==4.65.0
traitlets==5.14.1
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.4
yarg==0.1.9
zipp==3.15.0
